refactor(*): move Page to common types
