+++ src/components/app-header/types.ts
@@ -1,4 +1,4 @@
-import { Page } from '../app/types';
+import { Page } from '../../types/types';
 
 export type AppHeaderProps = {
     activePage: Page;
