+++ src/components/app/app.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import { AppContent } from '../app-content';
 import { AppHeader } from '../app-header';
 import './app.module.css';
-import { Page } from './types';
+import { Page } from '../../types/types';
 
 function App() {
     const [activePage, setActivePage] = useState<Page>(Page.CONSTRUCTOR);
