--- src/components/app/types.ts
@@ -1,5 +1,0 @@
-export enum Page {
-    CONSTRUCTOR = 'constructor',
-    ORDERS = 'orders',
-    ACCOUNT = 'account',
-}
